+++ obj/src/auth/index.js
@@ -6,4 +6,6 @@ var CredentialResolver_1 = require("./CredentialResolver");
 exports.CredentialResolver = CredentialResolver_1.CredentialResolver;
 var MemoryCredentialStore_1 = require("./MemoryCredentialStore");
 exports.MemoryCredentialStore = MemoryCredentialStore_1.MemoryCredentialStore;
+var DefaultCredentialStoreFactory_1 = require("./DefaultCredentialStoreFactory");
+exports.DefaultCredentialStoreFactory = DefaultCredentialStoreFactory_1.DefaultCredentialStoreFactory;
 //# sourceMappingURL=index.js.map