+++ obj/src/cache/DefaultCacheFactory.d.ts
@@ -1,9 +1,8 @@
-import { IFactory } from '../build/IFactory';
+import { Factory } from '../build/Factory';
 import { Descriptor } from '../refer/Descriptor';
-export declare class DefaultCacheFactory implements IFactory {
+export declare class DefaultCacheFactory extends Factory {
     static readonly Descriptor: Descriptor;
     static readonly NullCacheDescriptor: Descriptor;
     static readonly MemoryCacheDescriptor: Descriptor;
-    canCreate(locator: any): boolean;
-    create(locator: any): any;
+    constructor();
 }