+++ obj/src/cache/index.js
@@ -2,10 +2,10 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var CacheEntry_1 = require("./CacheEntry");
 exports.CacheEntry = CacheEntry_1.CacheEntry;
-var DefaultCacheFactory_1 = require("./DefaultCacheFactory");
-exports.DefaultCacheFactory = DefaultCacheFactory_1.DefaultCacheFactory;
 var NullCache_1 = require("./NullCache");
 exports.NullCache = NullCache_1.NullCache;
 var MemoryCache_1 = require("./MemoryCache");
 exports.MemoryCache = MemoryCache_1.MemoryCache;
+var DefaultCacheFactory_1 = require("./DefaultCacheFactory");
+exports.DefaultCacheFactory = DefaultCacheFactory_1.DefaultCacheFactory;
 //# sourceMappingURL=index.js.map