+++ obj/src/connect/index.js
@@ -6,4 +6,6 @@ var ConnectionResolver_1 = require("./ConnectionResolver");
 exports.ConnectionResolver = ConnectionResolver_1.ConnectionResolver;
 var MemoryDiscovery_1 = require("./MemoryDiscovery");
 exports.MemoryDiscovery = MemoryDiscovery_1.MemoryDiscovery;
+var DefaultDiscoveryFactory_1 = require("./DefaultDiscoveryFactory");
+exports.DefaultDiscoveryFactory = DefaultDiscoveryFactory_1.DefaultDiscoveryFactory;
 //# sourceMappingURL=index.js.map