+++ obj/src/count/DefaultCountersFactory.d.ts
@@ -1,10 +1,9 @@
-import { IFactory } from '../build/IFactory';
+import { Factory } from '../build/Factory';
 import { Descriptor } from '../refer/Descriptor';
-export declare class DefaultCountersFactory implements IFactory {
+export declare class DefaultCountersFactory extends Factory {
     static readonly Descriptor: Descriptor;
     static readonly NullCountersDescriptor: Descriptor;
     static readonly LogCountersDescriptor: Descriptor;
     static readonly CompositeCountersDescriptor: Descriptor;
-    canCreate(locator: any): boolean;
-    create(locator: any): any;
+    constructor();
 }