+++ obj/src/log/DefaultLoggerFactory.d.ts
@@ -1,10 +1,9 @@
-import { IFactory } from '../build/IFactory';
+import { Factory } from '../build/Factory';
 import { Descriptor } from '../refer/Descriptor';
-export declare class DefaultLoggerFactory implements IFactory {
+export declare class DefaultLoggerFactory extends Factory {
     static readonly Descriptor: Descriptor;
     static readonly NullLoggerDescriptor: Descriptor;
     static readonly ConsoleLoggerDescriptor: Descriptor;
     static readonly CompositeLoggerDescriptor: Descriptor;
-    canCreate(locator: any): boolean;
-    create(locator: any): any;
+    constructor();
 }