+++ src/cache/index.ts
@@ -1,5 +1,5 @@
 export { CacheEntry } from './CacheEntry';
 export { ICache } from './ICache';
-export { DefaultCacheFactory } from './DefaultCacheFactory';
 export { NullCache } from './NullCache';
 export { MemoryCache } from './MemoryCache';
+export { DefaultCacheFactory } from './DefaultCacheFactory';