+++ src/connect/index.ts
@@ -2,3 +2,4 @@ export { ConnectionParams } from './ConnectionParams';
 export { IDiscovery } from './IDiscovery';
 export { ConnectionResolver } from './ConnectionResolver';
 export { MemoryDiscovery } from './MemoryDiscovery';
+export { DefaultDiscoveryFactory } from './DefaultDiscoveryFactory';